Update lifx-group-of-groups.groovy
Changes to only update the on/off status of the group if all the bulbs are on. If all bulbs in the group are not on, the state remains off and the lightStatus shows the number of lights online vs always saying 'x of x'. Without looking at the status of all lights, the state really just depends on the state of the last bulb in the response from LIFX. 

Also fixed some typos.